+++ projects/06/Assembler/Global.h
@@ -8,5 +8,7 @@
 #include <iostream>
 #include <fstream>
 #include <string>
+#include <bitset>
+#include <map>
 
 #endif
